+++ toolkit/cli/smart-contracts-commands/src/governance.rs
@@ -1,4 +1,5 @@
 use crate::PaymentFilePath;
+use anyhow::anyhow;
 use partner_chains_cardano_offchain::{
 	await_tx::FixedDelayRetries, init_governance::run_init_governance,
 	update_governance::run_update_governance,
@@ -27,7 +28,7 @@ impl GovernanceCmd {
 pub struct InitGovernanceCmd {
 	#[clap(flatten)]
 	common_arguments: crate::CommonArguments,
-	/// Governance authority hash to be set.
+	/// Governance authority to be set
 	#[arg(long, short = 'g')]
 	governance_authority: MainchainKeyHash,
 	#[clap(flatten)]
@@ -58,9 +59,12 @@ impl InitGovernanceCmd {
 pub struct UpdateGovernanceCmd {
 	#[clap(flatten)]
 	common_arguments: crate::CommonArguments,
-	/// Governance authority hash to be set.
-	#[arg(long, short = 'g')]
-	new_governance_authority: MainchainKeyHash,
+	/// New governance authorities keys hashes, hex encoded, space delimited, order does not matter
+	#[arg(short = 'g', long, num_args = 1.., value_delimiter = ' ')]
+	new_governance_authority: Vec<MainchainKeyHash>,
+	/// Governance threshold to be set
+	#[arg(long)]
+	new_governance_threshold: u8,
 	#[clap(flatten)]
 	payment_key_file: PaymentFilePath,
 	/// Genesis UTXO of the chain
@@ -70,11 +74,19 @@ pub struct UpdateGovernanceCmd {
 
 impl UpdateGovernanceCmd {
 	pub async fn execute(self) -> crate::CmdResult<()> {
+		if self.new_governance_threshold > self.new_governance_authority.len() as u8 {
+			return Err(anyhow!(
+				"New governance threshold is greater than the number of governance authorities"
+			)
+			.into());
+		}
+
 		let payment_key = self.payment_key_file.read_key()?;
 		let client = self.common_arguments.get_ogmios_client().await?;
 
 		run_update_governance(
-			self.new_governance_authority,
+			&self.new_governance_authority,
+			self.new_governance_threshold,
 			&payment_key,
 			self.genesis_utxo,
 			&client,